fix lexer.c parser.c
