+++ Makefile
@@ -21,7 +21,6 @@ SRCS := $(SRCS_DIR)/main.c \
 		$(SRCS_DIR)/builtin/pwd.c \
 		$(SRCS_DIR)/builtin/unset.c \
 		$(SRCS_DIR)/lexer/lexer.c \
-		$(SRCS_DIR)/lexer/lexer_01.c \
 		$(SRCS_DIR)/astree/astree.c \
 		$(SRCS_DIR)/utils/x_func.c \
 		$(SRCS_DIR)/execute/execute.c \
