+++ includes/parser.h
@@ -1,22 +1,26 @@
 #ifndef PARSER_H
-#define PARSER_H
+# define PARSER_H
 
-#include "lexer.h"
-#include "minishell.h"
+
+# include <stdio.h>
+# include "lexer.h"
+# include "minishell.h"
+
+typedef struct s_cmd	t_cmd;
 
 typedef struct s_cmd
 {
-	char			*cmd;
-	t_list			*args;
-	//char			*stdin_str;
-	//t_list			*heredoc_end;
-	int				fd_out;
-	struct s_cmd	*prev;
-	struct s_cmd	*next;
+	char	*cmd;
+	t_list	*args;
+//	char			*stdin_str;
+//	t_list			*heredoc_end;
+	int		fd_out;
+//	t_cmd	*prev;
+	t_cmd	*next;
 }					t_cmd;
 
-t_cmd *parse(t_lexer *lexerbuf);
+t_cmd	*lex_pars(char *input);;
 void	cmd_add_back(t_cmd **lst, t_cmd *new);
 t_cmd	*cmd_new(char *cmd);
 
-#endif
+#endif // !PARSER_H
