+++ srcs/parser/parser.c
@@ -1,4 +1,5 @@
-#include "./parser.h"
+#include "parser.h"
+#include "lexer.h"
 
 void get_cmd_name(t_cmd *cmd_node, t_token **token)
 {
@@ -9,6 +10,7 @@ void get_cmd_name(t_cmd *cmd_node, t_token **token)
 	}
 	(*token) = (*token)->next;
 }
+
 void get_cmd_args(t_cmd *cmd, t_token **token)
 {
 	while ((*token) != NULL && strcmp((*token)->data, "|") != 0)
@@ -22,6 +24,8 @@ void set_cmd_info(t_cmd *list, t_token **token)
 {
 	while ((*token) != NULL)
 	{
+		if ((*token)->data == CHAR_PIPE);
+		
 		if (ft_strncmp((*token)->data, "|", 2) == 0)
 		{
 			*token = (*token)->next;
@@ -31,12 +35,12 @@ void set_cmd_info(t_cmd *list, t_token **token)
 		get_cmd_args(list, token);
 	}
 }
-t_cmd *init_t_cmd(t_lexer *lexerbuf)
+
+t_cmd *parse(t_lexer *lexerbuf)
 {
 	t_cmd *list;
 	t_cmd *new_node;
 	t_token *token;
-
 
 	list = NULL;
 	new_node = NULL;
@@ -51,14 +55,14 @@ t_cmd *init_t_cmd(t_lexer *lexerbuf)
 		set_cmd_info(new_node, &token);
 	}
 	return (list);
-
-
 }
-t_cmd *parse(t_lexer *lexerbuf)
-{
-	t_cmd *ret;
 
-	//lexerした値のvalidateをする
-	ret = init_t_cmd(lexerbuf);
-	return (ret);
+t_cmd	*lex_pars(char *input)
+{
+	t_cmd	*cmd_list;
+	t_lexer	lexerbuf;
+
+	lexer_build(input, &lexerbuf);
+	cmd_list = parse(&lexerbuf);
+	return (cmd_list);
 }
